requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
python-dotenv>=1.0.1
transformers>=4.43.0
torch>=2.3.0
PyMuPDF>=1.24.3
python-docx>=1.1.0
google-generativeai>=0.7.2
reportlab>=4.2.2
regex>=2024.5.15
